Add a Formula struct
I think it’s intuitive to have the Formula own all
the variables and clauses. However, clauses also
need to refer to variables which is hard to get
working according to Rust’s ownership rules.

Luckily, a variable is just a wrapper around a
number, so we can derive the Copy trait and then
clauses make copies of their variables, rather
than holding references to the same data.

It’s also nice to use sets to hold the variables
in clauses and the clauses in the formula because
they implicitly de-duplicate elements.

To use a type in a HashSet it needs to implement
Hash and HashSet (itself) does not. However,
BTreeSet does so we can combine the two.