Add a State struct
The index function takes the ‘name’ of a state and
returns the index at which this state would appear
in a lexicographically ordered list of state
names, e.g.

name: 12, index: 1
name: 13, index: 2
name: 21, index: 3
... etc

The way this is calculated is fairly complicated
and probably quite inefficient, but it’ll do for
now. These indexes could be precomputed then
memoized if needed.