Reverse the direction of implication for state transitions
false can imply true, which previously meant the
machine could transition to a state if the
conditions weren’t met.

The relation between states should probably be
if-and-only-if, but for the specific problem we’re
modelling, it should be sufficient to have
one-sided implication because it’s in the solver’s
‘best interest’ to choose to transition to states
whenever possible.

The advantage of this is we can omit lots of
clauses. I don’t know if it would help the solver
to include these clauses, but for now I’m going to
assume that fewer clauses is better.