Add a Goal struct
I’m trying an alternate approach to the
‘at least one’ clauses that are needed for the
final states. I think, instead, I can represent
each permutation as a ‘subgoal’ and then use a
similar approach of defining variables in a binary
string that are shared across states.

This will change the semantics from ‘at least one’
to ‘exactly one’ for each permutation to appear in
the string.

As far as I know, no one has proved that minimal
superpermutations don’t contain the same
permutation more than once, but it seems likely
and it’s probably the only hope of making the
problem tractable.