Replace ‘at_least_one’ clauses with Goal clauses
We’re now using binary to represent the location
in the string at which each permutation occurs.
This gives us ‘exactly one’ semantics which means
we can remove the expensive ‘at_least_one’ clauses
from the formula.

Currently, there’s nothing to stop the binary
representations for states taking on values that
don’t map to any states, so the formula is
satisfiable when it shouldn’t be. The next step is
to add clauses to prevent these invalid states.