Add Problem#each_permutation_appears_at_most_once
We no longer require that all permutations appear
in the string. Instead, we make it so each
permutation can appear at most once. If it does
appear, then the final state of the machine
implies the position it appears at, otherwise the
goal state is ‘dead’ and points to the start of
the string.